yt/requirements.txt
youtube-dl
python-vlc
pafy